+++ src/rapidfuzz/_utils.py
@@ -20,6 +20,7 @@ except BaseException:
 class ScorerFlag(IntFlag):
     RESULT_F64 = 1 << 5
     RESULT_I64 = 1 << 6
+    RESULT_SIZE_T = 1 << 7
     SYMMETRIC = 1 << 11
 
 
@@ -27,7 +28,7 @@ def _get_scorer_flags_distance(**_kwargs: Any) -> dict[str, Any]:
     return {
         "optimal_score": 0,
         "worst_score": 2**63 - 1,
-        "flags": ScorerFlag.RESULT_I64 | ScorerFlag.SYMMETRIC,
+        "flags": ScorerFlag.RESULT_SIZE_T | ScorerFlag.SYMMETRIC,
     }
 
 
@@ -35,7 +36,7 @@ def _get_scorer_flags_similarity(**_kwargs: Any) -> dict[str, Any]:
     return {
         "optimal_score": 2**63 - 1,
         "worst_score": 0,
-        "flags": ScorerFlag.RESULT_I64 | ScorerFlag.SYMMETRIC,
+        "flags": ScorerFlag.RESULT_SIZE_T | ScorerFlag.SYMMETRIC,
     }